Use QScintilla instead of own implementation of a code editor
This is a first rough implementation of QScintilla support as SQL text
editor.

It *should* work mostly and build fine with qmake as well as cmake.

As of now these features are broken:
- no separate colours for highlighting table names and function names
- no auto completion for table fields

However, even despite these problems I see the already as an improvement
as we have a much nicer auto completion now and there are plenty new
features which just need to be activated.